IK angles work! servoVals not yet
The IK works given the current position as hardcoded and produced the
expected angles for each joint. The conversion of angles to PWM values
for the servo still needs work.
Some function to input a goal position is obviously still missing.

+++ ik.h
@@ -23,9 +23,9 @@ Inverse kinematics
 
 /* cartesian position of end effector*/
 typedef struct {
-	int x_pos;
-	int y_pos;
-	int z_pos;
+	double x_pos;
+	double y_pos;
+	double z_pos;
 }ik_cart_pos_s;
